WIB-42 Rename | client : 경로 수정으로 import 구문 수정
